# src/docs/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "post": {
        "operationId": "ProductController_create",
        "summary": "",
        "description": "Creates a new product in the marketplace.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The product has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not found."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "products"
        ]
      },
      "get": {
        "operationId": "ProductController_findAll",
        "summary": "",
        "description": "Retrieves a list of all products.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of products retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductSchema"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductController_findOne",
        "summary": "",
        "description": "Retrieves a specific product by ID.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSchema"
                }
              }
            }
          },
          "404": {
            "description": "Product not found."
          }
        },
        "tags": [
          "products"
        ]
      },
      "put": {
        "operationId": "ProductController_update",
        "summary": "",
        "description": "Updates a specific product by ID.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSchema"
                }
              }
            }
          },
          "404": {
            "description": "Product not found."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductController_remove",
        "summary": "",
        "description": "Deletes a specific product by ID.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully."
          },
          "404": {
            "description": "Product not found."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "products"
        ]
      }
    },
...
  },
  "info": {
    "title": "Farmers Marketplace API",
    "description": "API documentation for the Farmers Marketplace platform",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    },
    {
      "url": "/api/v1"
    },
    {
      "url": "/api/v2"
    },
    {
      "url": "/api/v1"
    },
    {
      "url": "/api/v2"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
    "schemas": {
      "CreateProductDTO": {
        "type": "object",
        "properties": {}
      },
      "ProductSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the product"
          },
          "name": {
            "type": "string",
            "description": "Name of the product"
          },
          "description": {
            "type": "string",
            "description": "Description of the product",
            "nullable": true
          },
          "price": {
            "type": "number",
            "description": "Price of the product"
          },
          "unit": {
            "type": "string",
            "description": "Unit of measurement (e.g., kg, liter)"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of image URLs"
          },
          "stockQuantity": {
            "type": "number",
            "description": "Available stock quantity"
          },
          "harvestDate": {
            "type": "string",
            "format": "date-time",
            "description": "Date of harvest"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "name",
          "price",
          "unit",
          "stockQuantity",
          "harvestDate"
        ]
      },
      "UpdateProductDTO": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}